me and jim working on allTasks button

+++ app/api/task_routes.py
@@ -7,11 +7,32 @@ from app.forms.task_form import TaskForm
 task_routes = Blueprint('task', __name__)
 
 
-# GET Task
+# GET Single Task
 @task_routes.route('/')
 @login_required
 def get_task_info():
     taskId = request.json['taskId']
+    taskInfo = Task.query.get(taskId)
+
+    task = taskInfo.to_dict()
+    task['notes'] = {}
+
+    for note in taskInfo.taskNote:
+        task['notes'][note.id] = note.to_dict()
+        task['notes'][note.id]['username'] = {}
+        task['notes'][note.id]['username'] = note.noteUser.username
+
+    return {'task': task}
+
+# GET All Tasks
+@task_routes.route('/all')
+@login_required
+def get_all_tasks():
+    userId = current_user.id
+    tasks = Task.query.filter(Task.creatorId == userId).order_by(Task.createdAt).all()
+    #grab all tasks assigned to me 
+    return {"tasks": [task.to_dict() for task in tasks]}
+
 
     taskInfo = Task.query.get(taskId)
 
@@ -24,6 +45,7 @@ def get_task_info():
         task['notes'][note.id]['username'] = note.noteUser.username
 
     return {'task': task}
+
 
 
 # CREATE Task
@@ -53,50 +75,55 @@ def create_task():
 @task_routes.route('/', methods=['PATCH'])
 @login_required
 def update_task():
-    # newUserId for adding a user to a task
-    # can be 0 to signify no new user is being added
-    # to a task
-    newUserId = request.json['newUserId']
+    form = TaskForm()
+    form['csrf_token'].data = request.cookies['csrf_token']
+    if form.validate_on_submit():
 
-    # assignedUserId for determining if task owner
-    # would like to unassign task from currently
-    # assigned user
-    assignedUserId = request.json['assignedUserId']
+        # newUserId for adding a user to a task
+        # can be 0 to signify no new user is being added
+        # to a task
+        newUserId = request.json['newUserId']
 
-    userId = current_user.id
-    taskId = request.json['taskId']
-    content = request.json['content']
-    completed = request.json['completed']
-    startDate = request.json['startDate']
-    dueDate = request.json['dueDate']
-    priority = request.json['priority']
+        # assignedUserId for determining if task owner
+        # would like to unassign task from currently
+        # assigned user
+        assignedUserId = request.json['assignedUserId']
 
-    currentTask = Task.query.get(taskId)
+        userId = current_user.id
+        taskId = request.json['taskId']
+        content = request.json['content']
+        completed = request.json['completed']
+        startDate = request.json['startDate']
+        dueDate = request.json['dueDate']
+        priority = request.json['priority']
 
-    if content:
-        currentTask.content = content
-    if completed is not None:
-        currentTask.completed = completed
-    if startDate:
-        currentTask.startDate = startDate
-    if dueDate:
-        currentTask.dueDate = dueDate
-    if priority:
-        currentTask.priority = priority
-    if assignedUserId > 0 and newUserId == 0:
-        # unassigning task from currently assigned user
-        db.session.execute(f'''DELETE FROM givetousers
-        WHERE user_id={assignedUserId} AND task_id={taskId};''')
-    if assignedUserId > 0 and newUserId > 0:
-        # assigning task to a new user
-        db.session.execute(f'''INSERT INTO givetousers (user_id, task_id)
-        VALUES ({newUserId}, {taskId});''')
+        currentTask = Task.query.get(taskId)
 
-    currentTask.updatedAt = datetime.now()
+        if content:
+            currentTask.content = content
+        if completed is not None:
+            currentTask.completed = completed
+        if startDate:
+            currentTask.startDate = startDate
+        if dueDate:
+            currentTask.dueDate = dueDate
+        if priority:
+            currentTask.priority = priority
+        if assignedUserId > 0 and newUserId == 0:
+            # unassigning task from currently assigned user
+            db.session.execute(f'''DELETE FROM givetousers
+            WHERE user_id={assignedUserId} AND task_id={taskId};''')
+        if assignedUserId > 0 and newUserId > 0:
+            # assigning task to a new user
+            db.session.execute(f'''INSERT INTO givetousers (user_id, task_id)
+            VALUES ({newUserId}, {taskId});''')
 
-    db.session.commit()
+        currentTask.updatedAt = datetime.now()
 
-    return {'task': currentTask.to_dict()}
+        db.session.commit()
+
+        return {'task': currentTask.to_dict()}
+    return {'errors': validation_errors_to_error_messages(form.errors)}, 401
 
 
 # DELETE Task
